fix: swagger and sdk added
