+++ packages/sdk/HttpError.ts
