+++ packages/sdk/README.md
@@ -1,0 +1,38 @@
+# SDK Library
+This is a SDK library generated by [`nestia`](https://nestia.io).
+
+With this SDK library, you can easily and safely interact with backend server.
+
+Just import and call some API functions like gif image below:
+
+![nestia-sdk-demo](https://user-images.githubusercontent.com/13158709/215004990-368c589d-7101-404e-b81b-fbc936382f05.gif)
+
+> Left is server code, and right is client code utilizing the SDK
+
+
+
+
+# What [`Nestia`](https://nestia.io) is:
+![Nestia Logo](https://nestia.io/logo.png)
+
+[![GitHub license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/samchon/nestia/blob/master/LICENSE)
+[![npm version](https://img.shields.io/npm/v/@nestia/core.svg)](https://www.npmjs.com/package/@nestia/core)
+[![Downloads](https://img.shields.io/npm/dm/nestia.svg)](https://www.npmjs.com/package/nestia)
+[![Build Status](https://github.com/samchon/nestia/workflows/build/badge.svg)](https://github.com/samchon/nestia/actions?query=workflow%3Abuild)
+[![Guide Documents](https://img.shields.io/badge/guide-documents-forestgreen)](https://nestia.io/docs/)
+
+[Nestia](https://nestia.io) is a set of helper libraries for NestJS, supporting below features:
+
+  - `@nestia/core`: super-fast decorators
+  - `@nestia/sdk`
+    - SDK generator for clients
+    - Swagger generator evolved than ever
+    - Automatic E2E test functions generator
+  - `nestia`: just CLI (command line interface) tool
+
+> **Note**
+> 
+> - **Only one line** required, with pure TypeScript type
+> - Runtime validator is **20,000x faster** than `class-validator`
+> - JSON serialization is **200x faster** than `class-transformer`
+> - SDK is similar with [tRPC](https://trpc.io), but much advanced
