+++ packages/sdk/functional/prizes/index.ts
@@ -1,7 +1,7 @@
 /**
  * @packageDocumentation
  * @module api.functional.prizes
- * @nestia Generated by Nestia - https://github.com/samchon/nestia
+ * @nestia Generated by Nestia - https://github.com/samchon/nestia 
  */
 //================================================================
 export * as proposals from "./proposals";
