+++ packages/sdk/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@ORGANIZATION/PROJECT-api",
+  "version": "0.1.0",
+  "description": "SDK library generated by Nestia",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "scripts": {
+    "build": "npm run build:sdk && npm run compile",
+    "build:sdk": "rimraf /functional && cd ../../backend && npx nestia sdk && cd ../packages/sdk",
+    "compile": "rimraf lib && tsc",
+    "deploy": "npm run build && npm publish"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/samchon/nestia"
+  },
+  "author": "Jeongho Nam",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/samchon/nestia/issues"
+  },
+  "homepage": "https://nestia.io",
+  "files": [
+    "lib",
+    "package.json",
+    "README.md"
+  ]
+}
